+++ src/davinci_crawling/proxy/tests/test_proxy_mesh.py
@@ -4,6 +4,7 @@ import logging
 
 from caravaggio_rest_api.tests import CaravaggioBaseTest
 from davinci_crawling.proxy.proxy_mesh import ProxyMesh
+from django.conf import settings
 
 _logger = logging.getLogger("davinci_crawling.testing")
 
@@ -27,6 +28,37 @@ class TestProxyMesh(CaravaggioBaseTest):
         proxy_mesh = ProxyMesh()
         proxies = proxy_mesh.get_to_use_proxies()
         self.assertTrue(len(proxies) > 0)
+
+    def test_available_proxies_with_country_restrictions(self):
+        """
+        Test if we can get the available proxies from the proxy mesh api,
+        with country restrictions on the settings.
+        """
+        proxy_mesh_settings = settings.DAVINCI_CONF["architecture-params"][
+            "proxy"]["proxy_mesh"]
+
+        self._assert_proxies_are_from(proxy_mesh_settings, ["fr"])
+        self._assert_proxies_are_from(proxy_mesh_settings, ["us"])
+        self._assert_proxies_are_from(proxy_mesh_settings, ["fr", "us"])
+
+    def _assert_proxies_are_from(self, proxy_mesh_settings, countries):
+        countries_str = ",".join(countries)
+
+        proxy_mesh_settings["only-proxies-from"] = countries_str
+        ProxyMesh.to_use_proxies = None
+        ProxyMesh.available_proxies = None
+        proxy_mesh = ProxyMesh()
+        proxies = proxy_mesh.get_to_use_proxies()
+        for proxy in proxies:
+            proxy_http = proxy["http"]
+            proxy_http = proxy_http.split("@")[1]
+            proxy_country = proxy_http[0:2]
+
+            for country in countries:
+                if country == proxy_country:
+                    break
+            else:
+                self.fail("Returned an invalid country")
 
     def test_get_proxy_address(self):
         """